+++ packages/storeadmin/pages/Manufacturer/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from "next";
 import List from "../../components/manufacturer/index";
 const ManufacturerLists: NextPage = () => {
   return (
-    <main className=" ms-sm-autopx-md-4">
+    <main className=" ms-sm-auto px-md-4">
       <List />
     </main>
   );
