--- packages/storeadmin/pages/index.tsx
@@ -1,8 +1,0 @@
-import type { NextPage } from 'next';
-import Dashboard from '@/components/dashboard/index';
-
-const Home: NextPage = () => {
-  return <Dashboard />;
-};
-
-export default Home;
